server/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
pytest
requests
psycopg2-binary
python-keycloak
python-jose